pyproject.toml
[project]
name = "chainlit-realtime"
version = "0.1.0"
description = "Chainlit image analysis app (Vision API, OpenAI-compatible)"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv==1.0.1",
    "chainlit==2.6.2",
    "aiohttp>=3.11.11,<4",
    "pillow==10.4.0",
    "openai>=1.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
